+++ src/main/java/com/dpw/runner/shipment/services/service/v1/impl/V1ServiceImpl.java
@@ -828,7 +828,7 @@ public class V1ServiceImpl implements IV1Service {
             long time = System.currentTimeMillis();
             HttpEntity<Object> entity = new HttpEntity<>(request, V1AuthHelper.getHeaders());
             masterDataResponse = this.restTemplate.postForEntity(this.VESSEL_DATA_URL, entity, V1DataResponse.class, new Object[0]);
-            double timeTaken = System.currentTimeMillis() - time;
+            double timeTaken = (double) System.currentTimeMillis() - time;
             log.info("Token time taken in getVesselData() function {} with Request ID: {}", timeTaken, LoggerHelper.getRequestIdFromMDC());
             return masterDataResponse.getBody();
         } catch (HttpClientErrorException | HttpServerErrorException ex) {
